requirements.txt
Bottleneck==1.3.4
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.4
cloudpickle==2.0.0
cryptography==36.0.0
cycler==0.11.0
cytoolz==0.11.0
dask==2022.2.1
fonttools==4.25.0
fsspec==2022.2.0
idna==3.3
imagecodecs==2021.8.26
imageio==2.9.0
kiwisolver==1.3.2
locket==0.2.1
lxml==4.8.0
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
networkx==2.7.1
numexpr==2.8.1
numpy==1.21.5
packaging==21.3
pandas==1.4.2
partd==1.2.0
Pillow==9.0.1
pip==21.2.4
progress==1.6
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.4
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
scikit-image==0.19.2
scipy==1.7.3
setuptools==61.2.0
sip==4.19.13
six==1.16.0
tifffile==2021.7.2
toolz==0.11.2
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tornado==6.1
typing_extensions==4.1.1
urllib3==1.26.9
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2
